Support for flag parameters in comments of a flag
This commit introduces the ability to add a parameter describtion for a flag, so it will appear in the in the syntax of the doc of the cli.
This way you can write clis that gives better hinting of how to use the cli based app.

Parameters are marked with '@param' following the Javadoc-styled documentation standard.
The following example has a flag called 'path' that expects a directory name

class MyFlags {
 /**
        @param <directory>
        Search in 'directory'
 **/
@:flag('path')
 public var path:String = null;
 .
 .
 .
}

Cli.getDoc(new MyFlags(), new tink.cli.doc.DefaultFormatter() will then give this output:

Usage:
 --path, -p<directory> : Search in path of 'directory'

- Only the first occurence of @param is used since a flag only accepts one value,
- Lines with @params are removed from the comments before the are added to the doc of the flag.